[BNR-357] feat(diffusion): add the atom diffusion flow
The atom-level diffusion loop end to end: the layers and modules it
needs, wired into the Boltz models and their examples.

GitLab-MR: !52
GitOrigin-RevId: 441513ef6cc6f9a57a87e58a9ec5dbc07d072c85

+++ tensorrt_bionemo/_torch/layers/sequence_local_atom.py
@@ -80,6 +80,8 @@ def query_to_keys(query: torch.Tensor,
     else:
         raise ValueError("Query tensor must be 3, 4, or 5 dimensions")
     # 2*K: number of areas, W//2: area size
+
     query = query.view(B, multiplicity, 2 * K, W // 2, D)
     return torch.einsum("b m j i d, j k -> b m k i d", query,
-                        keys_indexing_matrix).reshape(B, multiplicity, K, H, D)
+                        keys_indexing_matrix.to(query.dtype)).reshape(
+                            B, multiplicity, K, H, D)
